+++ CHANGELOG.md
@@ -1,1 +1,6 @@
+## [unreleased]
+
+### 🚜 Refactor
+
+- *(context)* Unify useContext and remove useMountedContext
 ## [0.44.3] - 2025-09-13